feat(OpenCV): Create the template matcher (#2)
Create the Template matching class for locating the area of the
prayer points and extracting it. After the extration, the picture is
scaled up 10 X in both X and Y axis. A gaussian blur is used. Then
A binary thresholding is used to convert the image into black and
white. After this, the image is written to disk in the "filter" dir.

[QA]:
Has tests: Yes
Reviewed by: Nil